Seprate out serializeVNode test into a separate file
