+++ debug/test/browser/debug.test.js
@@ -4,10 +4,8 @@ import {
 	teardown,
 	serializeHtml
 } from '../../../test/_util/helpers';
-import { serializeVNode, initDebug } from '../../src/debug'; // TODO: Can we somehow convert this? Perhaps move seralizeVNode tests into their own file?
+import 'preact/debug';
 import * as PropTypes from 'prop-types';
-
-initDebug();
 
 const h = createElement;
 /** @jsx createElement */
@@ -317,66 +315,6 @@ describe('debug', () => {
 
 			render(<App />, scratch);
 			expect(console.error).to.be.calledTwice;
-		});
-	});
-
-	describe('serializeVNode', () => {
-		it("should prefer a function component's displayName", () => {
-			function Foo() {
-				return <div />;
-			}
-			Foo.displayName = 'Bar';
-
-			expect(serializeVNode(<Foo />)).to.equal('<Bar />');
-		});
-
-		it("should prefer a class component's displayName", () => {
-			class Bar extends Component {
-				render() {
-					return <div />;
-				}
-			}
-			Bar.displayName = 'Foo';
-
-			expect(serializeVNode(<Bar />)).to.equal('<Foo />');
-		});
-
-		it('should serialize vnodes without children', () => {
-			expect(serializeVNode(<br />)).to.equal('<br />');
-		});
-
-		it('should serialize vnodes with children', () => {
-			expect(serializeVNode(<div>Hello World</div>)).to.equal('<div>..</div>');
-		});
-
-		it('should serialize components', () => {
-			function Foo() {
-				return <div />;
-			}
-			expect(serializeVNode(<Foo />)).to.equal('<Foo />');
-		});
-
-		it('should serialize props', () => {
-			expect(serializeVNode(<div class="foo" />)).to.equal(
-				'<div class="foo" />'
-			);
-
-			let noop = () => {};
-			expect(serializeVNode(<div onClick={noop} />)).to.equal(
-				'<div onClick="function noop() {}" />'
-			);
-
-			function Foo(props) {
-				return props.foo;
-			}
-
-			expect(serializeVNode(<Foo foo={[1, 2, 3]} />)).to.equal(
-				'<Foo foo="1,2,3" />'
-			);
-
-			expect(serializeVNode(<div prop={Object.create(null)} />)).to.equal(
-				'<div prop="[object Object]" />'
-			);
 		});
 	});
